Created functions.py; Added Login Validation
I created a new file to store functions for the sake of organization and separation.

Also added to routes.py, admin.html, and functions.py functionality so that the user's entered credentials are validated against the passcodes file and an error is displayed if login details don't match with any records.

Paste files into the "flask_wtforms_tutorial" folder in place of the initial files the project starter code came with.